feat(hashivault): add GitHub Actions OIDC support and Vault token caching
Add a new `WERF_ACTIONS_AUDIENCE` environment variable that enables
GitHub Actions OIDC flow. When set, the client requests a fresh JWT
from the GitHub OIDC endpoint (`ACTIONS_ID_TOKEN_REQUEST_URL`) before
each Vault login, solving the 10-minute OIDC token expiry issue.

Also introduce Vault token caching across all authenticators (JWT and
AppRole): the obtained Vault token is cached and reused until it expires
(with a 30-second safety margin), avoiding unnecessary re-authentication
on every operation.
