+++ pkg/signature/image/manifest_vault_remote_test.go
@@ -1,0 +1,141 @@
+package image_test
+
+import (
+	"fmt"
+	"os"
+	"time"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+
+	"github.com/deckhouse/delivery-kit-sdk/pkg/signver"
+)
+
+var _ = Describe("manifest with real vault using remote Vault server", Label("e2e", "remote"), func() {
+	DescribeTable("sign and verify image manifest one time using static auth token with VAULT_TOKEN",
+		Label("vault-token"),
+		func(ctx SpecContext, keyRef, certRef, chainRef, rootRef string, useVerification bool) {
+			if os.Getenv("TRANSIT_SECRET_ENGINE_PATH") == "" ||
+				os.Getenv("VAULT_ADDR") == "" ||
+				os.Getenv("VAULT_TOKEN") == "" {
+				Skip("Skipped because TRANSIT_SECRET_ENGINE_PATH, VAULT_ADDR or VAULT_TOKEN is not set")
+			}
+
+			sv, err := signver.NewSignerVerifier(ctx, certRef, chainRef, signver.KeyOpts{
+				KeyRef: keyRef,
+			})
+			Expect(err).To(Succeed())
+
+			manifest := testSign(ctx, sv)
+
+			if !useVerification {
+				return
+			}
+
+			testVerify(ctx, manifest, []string{rootRef})
+		},
+		Entry(
+			"ED25519 key",
+			"hashivault://dh-2025-july",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUMyVENDQW91Z0F3SUJBZ0lVVUdLZDkyUkJGQ3dYL3Jla3ZsTE5taTRJd3pJd0JRWURLMlZ3TUhzeEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SSXdFQVlEVlFRTEV3bEVaV05yYUc5MWMyVXhJakFnQmdOVkJBTVRHVXBUClF5QkdiR0Z1ZENCSmJuUmxjbTFsWkdsaGRHVWdRMEV3SGhjTk1qVXdOekl4TVRNek5UQXhXaGNOTWpZd09ESXcKTVRNek5UTXhXakI0TVFzd0NRWURWUVFHRXdKU1ZURVBNQTBHQTFVRUNCTUdUVzl6WTI5M01ROHdEUVlEVlFRSApFd1pOYjNOamIzY3hFakFRQmdOVkJBb1RDVXBUUXlCR2JHRnVkREVTTUJBR0ExVUVDeE1KUkdWamEyaHZkWE5sCk1SOHdIUVlEVlFRRERCWmthQzF6YVdkdVpYSkFaR1ZqYTJodmRYTmxMbkoxTUNvd0JRWURLMlZ3QXlFQXpMWmQKajRtTnNvUkFFeFlmV3dWdExtdHpWWjRxekRIUmhVbVJDVEJrMDJDamdnRWlNSUlCSGpBT0JnTlZIUThCQWY4RQpCQU1DQTZnd0hRWURWUjBsQkJZd0ZBWUlLd1lCQlFVSEF3RUdDQ3NHQVFVRkJ3TUNNQjBHQTFVZERnUVdCQlQ2CkxubGtjMkRobythZFg0YjFjYk11VXBQSnBUQWZCZ05WSFNNRUdEQVdnQlFWSWdPNW83NjFqakE5VEZzaGhsVWUKR3duUDZ6QkpCZ2dyQmdFRkJRY0JBUVE5TURzd09RWUlLd1lCQlFVSE1BS0dMV2gwZEhCek9pOHZNVEkzTGpBdQpNQzR4T2pneU1EQXZkakV2Y0d0cExXWnNZVzUwTFdsdWRDMWthQzlqWVRBaEJnTlZIUkVFR2pBWWdSWmthQzF6CmFXZHVaWEpBWkdWamEyaHZkWE5sTG5KMU1EOEdBMVVkSHdRNE1EWXdOS0F5b0RDR0xtaDBkSEJ6T2k4dk1USTMKTGpBdU1DNHhPamd5TURBdmRqRXZjR3RwTFdac1lXNTBMV2x1ZEMxa2FDOWpjbXd3QlFZREsyVndBMEVBSGNseApCamF0aVVDOEN3SnNWalE4RUVrdHIwYVlQNnBNc0dDL3QrWkhMWUJSOVk3UmhxclFvb3lNVE5URXFnbE43REp5ClRuQ2I0U21yTEY0YnpmZDdDQT09Ci0tLS0tRU5EIENFUlRJRklDQVRFLS0tLS0=",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNvakNDQWxTZ0F3SUJBZ0lVWmNiNE9WTmhhbXJOTU53dlRFT1J5RGFJMHhjd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE16TXpVeldoY05NelV3TnpFNU1UTXoKTkRJeldqQjdNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVNNQkFHQTFVRUN4TUpSR1ZqYTJodmRYTmxNU0l3CklBWURWUVFERXhsS1UwTWdSbXhoYm5RZ1NXNTBaWEp0WldScFlYUmxJRU5CTUNvd0JRWURLMlZ3QXlFQTYyeG4KcVJEUm5uSGZDRXJlNFliYW55aDNDT2p6Z3N1alFSenJGSlVtS2Ztamdld3dnZWt3RGdZRFZSMFBBUUgvQkFRRApBZ0VHTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkJZaUI3bWp2cldPTUQxTVd5R0dWUjRiCkNjL3JNQjhHQTFVZEl3UVlNQmFBRkpCZ0s0NjYzVVdZMEhKbTVGb1plK1Rsem1mek1FY0dDQ3NHQVFVRkJ3RUIKQkRzd09UQTNCZ2dyQmdFRkJRY3dBb1lyYUhSMGNITTZMeTh4TWpjdU1DNHdMakU2T0RJd01DOTJNUzl3YTJrdApabXhoYm5RdGNtOXZkQzlqWVRBOUJnTlZIUjhFTmpBME1ES2dNS0F1aGl4b2RIUndjem92THpFeU55NHdMakF1Ck1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXliMjkwTDJOeWJEQUZCZ01yWlhBRFFRQnNOZWNUU0l6bmxvQWYKYkViU2JERlpGb20wMXc2WDN5WVpqVmFOQngrNEZmV3JNZ1ppU05rM3NkZ0dNRlpEVnVaZWl4OXlQUTk3bVIzWgpaTXU0Q3JRUAotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0t",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNGVENDQWNlZ0F3SUJBZ0lVZElhblV0Nk0vNnFTL0RzMXJWSjZoc09JTmYwd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE15TkRVMFdoY05ORFV3TnpFMk1UTXkKTlRJMFdqQjRNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVFNQTRHQTFVRUN4TUhTVzVtYjNObFl6RWhNQjhHCkExVUVBeE1ZU2xORElFWnNZVzUwSUZKdmIzUWdRMEVnUVMweU1ESTFNQ293QlFZREsyVndBeUVBbGd1Q3hQRHUKV1VhajZjazZHdFIrZGdCNS9SRENSNmdhSWkzN3ZCQmx0ZE9qWXpCaE1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUApCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFmCkJnTlZIU01FR0RBV2dCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFGQmdNclpYQURRUUJxbTd5S21mVG0KV01PcFdHRldxSVJLcmJnRW44NkpyQmVWSXFXakVUMWs2NEVjR2pFN0JkdWxqdjdDVGhCQ2xwY1c3bzFuUGlVeApZeDFUNnMvSmU3b0kKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			true,
+		),
+		Entry(
+			"ECDSA_P256 key",
+			"hashivault://dh-2025-aug-ec",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN0ekNDQW1tZ0F3SUJBZ0lVS25kUFo2aFl4Sm9pWW9kMkxmbVB4elFmcUNnd0JRWURLMlZ3TUhzeEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SSXdFQVlEVlFRTEV3bEVaV05yYUc5MWMyVXhJakFnQmdOVkJBTVRHVXBUClF5QkdiR0Z1ZENCSmJuUmxjbTFsWkdsaGRHVWdRMEV3SGhjTk1qVXdPREEzTURnMU1qRTVXaGNOTWpZd09UQTIKTURnMU1qUTVXakFrTVNJd0lBWURWUVFEREJsa2FDMXphV2R1WlhJdFpXTkFaR1ZqYTJodmRYTmxMbkoxTUZrdwpFd1lIS29aSXpqMENBUVlJS29aSXpqMERBUWNEUWdBRWlXRFNwWWRkcVBHY2h4SUIxZ1lTMmhwNVVKMkxMT1pHClpTYnNTSGdkVnRxZ2F2NE9WVGlKcHJVcFVXbjA4UG9LQndZOEdOU2dPaEhIZlU5NTYzV09TcU9DQVNVd2dnRWgKTUE0R0ExVWREd0VCL3dRRUF3SURxREFkQmdOVkhTVUVGakFVQmdnckJnRUZCUWNEQVFZSUt3WUJCUVVIQXdJdwpIUVlEVlIwT0JCWUVGSWRUK3RuS0ZmTjJ3YU1DYjV0ZGdITlJreEVNTUI4R0ExVWRJd1FZTUJhQUZCWWlCN21qCnZyV09NRDFNV3lHR1ZSNGJDYy9yTUVrR0NDc0dBUVVGQndFQkJEMHdPekE1QmdnckJnRUZCUWN3QW9ZdGFIUjAKY0hNNkx5OHhNamN1TUM0d0xqRTZPREl3TUM5Mk1TOXdhMmt0Wm14aGJuUXRhVzUwTFdSb0wyTmhNQ1FHQTFVZApFUVFkTUJ1QkdXUm9MWE5wWjI1bGNpMWxZMEJrWldOcmFHOTFjMlV1Y25Vd1B3WURWUjBmQkRnd05qQTBvREtnCk1JWXVhSFIwY0hNNkx5OHhNamN1TUM0d0xqRTZPREl3TUM5Mk1TOXdhMmt0Wm14aGJuUXRhVzUwTFdSb0wyTnkKYkRBRkJnTXJaWEFEUVFBRy8zQXlWMFJSS3lRSnFHT0ZOTU5GRTNiNzUrL3ViRGJwamVFaHhEcHNQMHB3ZU5abApmVUxqRmgwSVBob2k0TlB5L0pYYVVidWNycG45K2RWcDRHQUQKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNvakNDQWxTZ0F3SUJBZ0lVWmNiNE9WTmhhbXJOTU53dlRFT1J5RGFJMHhjd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE16TXpVeldoY05NelV3TnpFNU1UTXoKTkRJeldqQjdNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVNNQkFHQTFVRUN4TUpSR1ZqYTJodmRYTmxNU0l3CklBWURWUVFERXhsS1UwTWdSbXhoYm5RZ1NXNTBaWEp0WldScFlYUmxJRU5CTUNvd0JRWURLMlZ3QXlFQTYyeG4KcVJEUm5uSGZDRXJlNFliYW55aDNDT2p6Z3N1alFSenJGSlVtS2Ztamdld3dnZWt3RGdZRFZSMFBBUUgvQkFRRApBZ0VHTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkJZaUI3bWp2cldPTUQxTVd5R0dWUjRiCkNjL3JNQjhHQTFVZEl3UVlNQmFBRkpCZ0s0NjYzVVdZMEhKbTVGb1plK1Rsem1mek1FY0dDQ3NHQVFVRkJ3RUIKQkRzd09UQTNCZ2dyQmdFRkJRY3dBb1lyYUhSMGNITTZMeTh4TWpjdU1DNHdMakU2T0RJd01DOTJNUzl3YTJrdApabXhoYm5RdGNtOXZkQzlqWVRBOUJnTlZIUjhFTmpBME1ES2dNS0F1aGl4b2RIUndjem92THpFeU55NHdMakF1Ck1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXliMjkwTDJOeWJEQUZCZ01yWlhBRFFRQnNOZWNUU0l6bmxvQWYKYkViU2JERlpGb20wMXc2WDN5WVpqVmFOQngrNEZmV3JNZ1ppU05rM3NkZ0dNRlpEVnVaZWl4OXlQUTk3bVIzWgpaTXU0Q3JRUAotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0t",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNGVENDQWNlZ0F3SUJBZ0lVZElhblV0Nk0vNnFTL0RzMXJWSjZoc09JTmYwd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE15TkRVMFdoY05ORFV3TnpFMk1UTXkKTlRJMFdqQjRNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVFNQTRHQTFVRUN4TUhTVzVtYjNObFl6RWhNQjhHCkExVUVBeE1ZU2xORElFWnNZVzUwSUZKdmIzUWdRMEVnUVMweU1ESTFNQ293QlFZREsyVndBeUVBbGd1Q3hQRHUKV1VhajZjazZHdFIrZGdCNS9SRENSNmdhSWkzN3ZCQmx0ZE9qWXpCaE1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUApCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFmCkJnTlZIU01FR0RBV2dCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFGQmdNclpYQURRUUJxbTd5S21mVG0KV01PcFdHRldxSVJLcmJnRW44NkpyQmVWSXFXakVUMWs2NEVjR2pFN0JkdWxqdjdDVGhCQ2xwY1c3bzFuUGlVeApZeDFUNnMvSmU3b0kKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			true,
+		),
+	)
+
+	// ACTIONS_ID_TOKEN is a static JWT obtained from the GitHub OIDC endpoint once before the test run.
+	// Its maximum lifetime is 10 minutes in GitHub Actions.
+	DescribeTable("sign and verify image manifest one time using per request auth with WERF_VAULT_AUTH_JWT on GitHub CI",
+		Label("static-github-jwt"),
+		func(ctx SpecContext, keyRef, certRef, chainRef, rootRef string, useVerification bool) {
+			if os.Getenv("CI") != "true" ||
+				os.Getenv("VAULT_ADDR") == "" ||
+				os.Getenv("WERF_VAULT_AUTH_ROLE") == "" ||
+				os.Getenv("ACTIONS_ID_TOKEN") == "" {
+				Skip("Skipped because CI=false or VAULT_ADDR, WERF_VAULT_AUTH_ROLE or ACTIONS_ID_TOKEN is not set")
+			}
+
+			// GinkgoT().Setenv("VAULT_ADDR", "")
+			GinkgoT().Setenv("TRANSIT_SECRET_ENGINE_PATH", "dh-signer-dev")
+			GinkgoT().Setenv("WERF_VAULT_AUTH_PATH", "github")
+			GinkgoT().Setenv("WERF_VAULT_AUTH_JWT", os.Getenv("ACTIONS_ID_TOKEN"))
+
+			sv, err := signver.NewSignerVerifier(ctx, certRef, chainRef, signver.KeyOpts{
+				KeyRef: keyRef,
+			})
+			Expect(err).To(Succeed())
+
+			manifest := testSign(ctx, sv)
+
+			if !useVerification {
+				return
+			}
+
+			testVerify(ctx, manifest, []string{rootRef})
+		},
+		Entry(
+			"ED25519 key",
+			"hashivault://dh-2025-aug-ec",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN2ekNDQW5HZ0F3SUJBZ0lVSjc4SXVEaFVjM1NSZ1YwcVZUL2FKcGZzbHQwd0JRWURLMlZ3TUhzeEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SSXdFQVlEVlFRTEV3bEVaV05yYUc5MWMyVXhJakFnQmdOVkJBTVRHVXBUClF5QkdiR0Z1ZENCSmJuUmxjbTFsWkdsaGRHVWdRMEV3SGhjTk1qVXdPREkyTURjeU5UQXhXaGNOTWpZd09USTEKTURjeU5UTXhXakFvTVNZd0pBWURWUVFEREIxa2FDMXphV2R1WlhJdFpHVjJMV1ZqUUdSbFkydG9iM1Z6WlM1eQpkVEJaTUJNR0J5cUdTTTQ5QWdFR0NDcUdTTTQ5QXdFSEEwSUFCQ2x3Sm9oNzBLSTRSa2VYd3FGTFpCUTVIWUFHCm1jQzV0RzMyaHpDTUQ1Y244SFR0NDEvVDhxVXlpNGFvWjkvMmdVdXlGRzAxSFhVV2xCakFxOU5FRTQramdnRXAKTUlJQkpUQU9CZ05WSFE4QkFmOEVCQU1DQTZnd0hRWURWUjBsQkJZd0ZBWUlLd1lCQlFVSEF3RUdDQ3NHQVFVRgpCd01DTUIwR0ExVWREZ1FXQkJRZ0oySXUxbWlyaWhPUXpjRnpZU3hkT3RZMmZqQWZCZ05WSFNNRUdEQVdnQlFXCklnZTVvNzYxampBOVRGc2hobFVlR3duUDZ6QkpCZ2dyQmdFRkJRY0JBUVE5TURzd09RWUlLd1lCQlFVSE1BS0cKTFdoMGRIQnpPaTh2TVRJM0xqQXVNQzR4T2pneU1EQXZkakV2Y0d0cExXWnNZVzUwTFdsdWRDMWthQzlqWVRBbwpCZ05WSFJFRUlUQWZnUjFrYUMxemFXZHVaWEl0WkdWMkxXVmpRR1JsWTJ0b2IzVnpaUzV5ZFRBL0JnTlZIUjhFCk9EQTJNRFNnTXFBd2hpNW9kSFJ3Y3pvdkx6RXlOeTR3TGpBdU1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXAKYm5RdFpHZ3ZZM0pzTUFVR0F5dGxjQU5CQUREZ0tXOTZIZS9EN2drU3UvaFFXUExMTWtwR3pITytuY09KaUZxbwo5dGdtYTRGVkJvZU1RRDFGaDlkem9sRG1vSkZ5L1VwL2hld1EwN1JLeVhjNmNBUT0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQ==",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNvakNDQWxTZ0F3SUJBZ0lVWmNiNE9WTmhhbXJOTU53dlRFT1J5RGFJMHhjd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE16TXpVeldoY05NelV3TnpFNU1UTXoKTkRJeldqQjdNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVNNQkFHQTFVRUN4TUpSR1ZqYTJodmRYTmxNU0l3CklBWURWUVFERXhsS1UwTWdSbXhoYm5RZ1NXNTBaWEp0WldScFlYUmxJRU5CTUNvd0JRWURLMlZ3QXlFQTYyeG4KcVJEUm5uSGZDRXJlNFliYW55aDNDT2p6Z3N1alFSenJGSlVtS2Ztamdld3dnZWt3RGdZRFZSMFBBUUgvQkFRRApBZ0VHTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkJZaUI3bWp2cldPTUQxTVd5R0dWUjRiCkNjL3JNQjhHQTFVZEl3UVlNQmFBRkpCZ0s0NjYzVVdZMEhKbTVGb1plK1Rsem1mek1FY0dDQ3NHQVFVRkJ3RUIKQkRzd09UQTNCZ2dyQmdFRkJRY3dBb1lyYUhSMGNITTZMeTh4TWpjdU1DNHdMakU2T0RJd01DOTJNUzl3YTJrdApabXhoYm5RdGNtOXZkQzlqWVRBOUJnTlZIUjhFTmpBME1ES2dNS0F1aGl4b2RIUndjem92THpFeU55NHdMakF1Ck1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXliMjkwTDJOeWJEQUZCZ01yWlhBRFFRQnNOZWNUU0l6bmxvQWYKYkViU2JERlpGb20wMXc2WDN5WVpqVmFOQngrNEZmV3JNZ1ppU05rM3NkZ0dNRlpEVnVaZWl4OXlQUTk3bVIzWgpaTXU0Q3JRUAotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0t",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNGVENDQWNlZ0F3SUJBZ0lVZElhblV0Nk0vNnFTL0RzMXJWSjZoc09JTmYwd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE15TkRVMFdoY05ORFV3TnpFMk1UTXkKTlRJMFdqQjRNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVFNQTRHQTFVRUN4TUhTVzVtYjNObFl6RWhNQjhHCkExVUVBeE1ZU2xORElFWnNZVzUwSUZKdmIzUWdRMEVnUVMweU1ESTFNQ293QlFZREsyVndBeUVBbGd1Q3hQRHUKV1VhajZjazZHdFIrZGdCNS9SRENSNmdhSWkzN3ZCQmx0ZE9qWXpCaE1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUApCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFmCkJnTlZIU01FR0RBV2dCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFGQmdNclpYQURRUUJxbTd5S21mVG0KV01PcFdHRldxSVJLcmJnRW44NkpyQmVWSXFXakVUMWs2NEVjR2pFN0JkdWxqdjdDVGhCQ2xwY1c3bzFuUGlVeApZeDFUNnMvSmU3b0kKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			true,
+		),
+		Entry(
+			"ECDSA_P256 key",
+			"hashivault://dh-2025-aug-ec",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN2ekNDQW5HZ0F3SUJBZ0lVSjc4SXVEaFVjM1NSZ1YwcVZUL2FKcGZzbHQwd0JRWURLMlZ3TUhzeEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SSXdFQVlEVlFRTEV3bEVaV05yYUc5MWMyVXhJakFnQmdOVkJBTVRHVXBUClF5QkdiR0Z1ZENCSmJuUmxjbTFsWkdsaGRHVWdRMEV3SGhjTk1qVXdPREkyTURjeU5UQXhXaGNOTWpZd09USTEKTURjeU5UTXhXakFvTVNZd0pBWURWUVFEREIxa2FDMXphV2R1WlhJdFpHVjJMV1ZqUUdSbFkydG9iM1Z6WlM1eQpkVEJaTUJNR0J5cUdTTTQ5QWdFR0NDcUdTTTQ5QXdFSEEwSUFCQ2x3Sm9oNzBLSTRSa2VYd3FGTFpCUTVIWUFHCm1jQzV0RzMyaHpDTUQ1Y244SFR0NDEvVDhxVXlpNGFvWjkvMmdVdXlGRzAxSFhVV2xCakFxOU5FRTQramdnRXAKTUlJQkpUQU9CZ05WSFE4QkFmOEVCQU1DQTZnd0hRWURWUjBsQkJZd0ZBWUlLd1lCQlFVSEF3RUdDQ3NHQVFVRgpCd01DTUIwR0ExVWREZ1FXQkJRZ0oySXUxbWlyaWhPUXpjRnpZU3hkT3RZMmZqQWZCZ05WSFNNRUdEQVdnQlFXCklnZTVvNzYxampBOVRGc2hobFVlR3duUDZ6QkpCZ2dyQmdFRkJRY0JBUVE5TURzd09RWUlLd1lCQlFVSE1BS0cKTFdoMGRIQnpPaTh2TVRJM0xqQXVNQzR4T2pneU1EQXZkakV2Y0d0cExXWnNZVzUwTFdsdWRDMWthQzlqWVRBbwpCZ05WSFJFRUlUQWZnUjFrYUMxemFXZHVaWEl0WkdWMkxXVmpRR1JsWTJ0b2IzVnpaUzV5ZFRBL0JnTlZIUjhFCk9EQTJNRFNnTXFBd2hpNW9kSFJ3Y3pvdkx6RXlOeTR3TGpBdU1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXAKYm5RdFpHZ3ZZM0pzTUFVR0F5dGxjQU5CQUREZ0tXOTZIZS9EN2drU3UvaFFXUExMTWtwR3pITytuY09KaUZxbwo5dGdtYTRGVkJvZU1RRDFGaDlkem9sRG1vSkZ5L1VwL2hld1EwN1JLeVhjNmNBUT0KLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQ==",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNvakNDQWxTZ0F3SUJBZ0lVWmNiNE9WTmhhbXJOTU53dlRFT1J5RGFJMHhjd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE16TXpVeldoY05NelV3TnpFNU1UTXoKTkRJeldqQjdNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVNNQkFHQTFVRUN4TUpSR1ZqYTJodmRYTmxNU0l3CklBWURWUVFERXhsS1UwTWdSbXhoYm5RZ1NXNTBaWEp0WldScFlYUmxJRU5CTUNvd0JRWURLMlZ3QXlFQTYyeG4KcVJEUm5uSGZDRXJlNFliYW55aDNDT2p6Z3N1alFSenJGSlVtS2Ztamdld3dnZWt3RGdZRFZSMFBBUUgvQkFRRApBZ0VHTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkJZaUI3bWp2cldPTUQxTVd5R0dWUjRiCkNjL3JNQjhHQTFVZEl3UVlNQmFBRkpCZ0s0NjYzVVdZMEhKbTVGb1plK1Rsem1mek1FY0dDQ3NHQVFVRkJ3RUIKQkRzd09UQTNCZ2dyQmdFRkJRY3dBb1lyYUhSMGNITTZMeTh4TWpjdU1DNHdMakU2T0RJd01DOTJNUzl3YTJrdApabXhoYm5RdGNtOXZkQzlqWVRBOUJnTlZIUjhFTmpBME1ES2dNS0F1aGl4b2RIUndjem92THpFeU55NHdMakF1Ck1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXliMjkwTDJOeWJEQUZCZ01yWlhBRFFRQnNOZWNUU0l6bmxvQWYKYkViU2JERlpGb20wMXc2WDN5WVpqVmFOQngrNEZmV3JNZ1ppU05rM3NkZ0dNRlpEVnVaZWl4OXlQUTk3bVIzWgpaTXU0Q3JRUAotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0t",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNGVENDQWNlZ0F3SUJBZ0lVZElhblV0Nk0vNnFTL0RzMXJWSjZoc09JTmYwd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE15TkRVMFdoY05ORFV3TnpFMk1UTXkKTlRJMFdqQjRNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVFNQTRHQTFVRUN4TUhTVzVtYjNObFl6RWhNQjhHCkExVUVBeE1ZU2xORElFWnNZVzUwSUZKdmIzUWdRMEVnUVMweU1ESTFNQ293QlFZREsyVndBeUVBbGd1Q3hQRHUKV1VhajZjazZHdFIrZGdCNS9SRENSNmdhSWkzN3ZCQmx0ZE9qWXpCaE1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUApCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFmCkJnTlZIU01FR0RBV2dCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFGQmdNclpYQURRUUJxbTd5S21mVG0KV01PcFdHRldxSVJLcmJnRW44NkpyQmVWSXFXakVUMWs2NEVjR2pFN0JkdWxqdjdDVGhCQ2xwY1c3bzFuUGlVeApZeDFUNnMvSmU3b0kKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			true,
+		),
+	)
+
+	DescribeTable("sign image manifest multiple times using with APP_ROLE",
+		Label("approle"),
+		func(ctx SpecContext, keyRef, certRef, chainRef, rootRef string, signingExperimentTimeout, signingExperiment, signingAttemptInterval time.Duration) {
+			if os.Getenv("TRANSIT_SECRET_ENGINE_PATH") == "" ||
+				os.Getenv("VAULT_ADDR") == "" ||
+				os.Getenv("WERF_VAULT_AUTH_ROLE_ID") == "" ||
+				os.Getenv("WERF_VAULT_AUTH_SECRET_ID") == "" {
+				Skip("Skipped because TRANSIT_SECRET_ENGINE_PATH, VAULT_ADDR, WERF_VAULT_AUTH_ROLE_ID or WERF_VAULT_AUTH_SECRET_ID is not set")
+			}
+
+			sv, err := signver.NewSignerVerifier(ctx, certRef, chainRef, signver.KeyOpts{
+				KeyRef: keyRef,
+			})
+			Expect(err).To(Succeed())
+
+			experimentStartedAt := time.Now()
+			attempt := 1
+
+			Eventually(func() time.Duration {
+				attemptStartedAt := time.Now()
+				fmt.Printf("Signing %d ... ", attempt)
+				testSign(ctx, sv)
+				fmt.Printf("done (%s, %s).\n", time.Since(attemptStartedAt), time.Now().Format(time.TimeOnly))
+				attempt++
+				return time.Since(experimentStartedAt)
+			}, signingExperimentTimeout, signingAttemptInterval).Should(BeNumerically(">=", signingExperiment))
+		},
+		Entry(
+			"should operate with auth token which has ttl=3m",
+			"hashivault://dh-2025-aug-ec",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUN0ekNDQW1tZ0F3SUJBZ0lVS25kUFo2aFl4Sm9pWW9kMkxmbVB4elFmcUNnd0JRWURLMlZ3TUhzeEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SSXdFQVlEVlFRTEV3bEVaV05yYUc5MWMyVXhJakFnQmdOVkJBTVRHVXBUClF5QkdiR0Z1ZENCSmJuUmxjbTFsWkdsaGRHVWdRMEV3SGhjTk1qVXdPREEzTURnMU1qRTVXaGNOTWpZd09UQTIKTURnMU1qUTVXakFrTVNJd0lBWURWUVFEREJsa2FDMXphV2R1WlhJdFpXTkFaR1ZqYTJodmRYTmxMbkoxTUZrdwpFd1lIS29aSXpqMENBUVlJS29aSXpqMERBUWNEUWdBRWlXRFNwWWRkcVBHY2h4SUIxZ1lTMmhwNVVKMkxMT1pHClpTYnNTSGdkVnRxZ2F2NE9WVGlKcHJVcFVXbjA4UG9LQndZOEdOU2dPaEhIZlU5NTYzV09TcU9DQVNVd2dnRWgKTUE0R0ExVWREd0VCL3dRRUF3SURxREFkQmdOVkhTVUVGakFVQmdnckJnRUZCUWNEQVFZSUt3WUJCUVVIQXdJdwpIUVlEVlIwT0JCWUVGSWRUK3RuS0ZmTjJ3YU1DYjV0ZGdITlJreEVNTUI4R0ExVWRJd1FZTUJhQUZCWWlCN21qCnZyV09NRDFNV3lHR1ZSNGJDYy9yTUVrR0NDc0dBUVVGQndFQkJEMHdPekE1QmdnckJnRUZCUWN3QW9ZdGFIUjAKY0hNNkx5OHhNamN1TUM0d0xqRTZPREl3TUM5Mk1TOXdhMmt0Wm14aGJuUXRhVzUwTFdSb0wyTmhNQ1FHQTFVZApFUVFkTUJ1QkdXUm9MWE5wWjI1bGNpMWxZMEJrWldOcmFHOTFjMlV1Y25Vd1B3WURWUjBmQkRnd05qQTBvREtnCk1JWXVhSFIwY0hNNkx5OHhNamN1TUM0d0xqRTZPREl3TUM5Mk1TOXdhMmt0Wm14aGJuUXRhVzUwTFdSb0wyTnkKYkRBRkJnTXJaWEFEUVFBRy8zQXlWMFJSS3lRSnFHT0ZOTU5GRTNiNzUrL3ViRGJwamVFaHhEcHNQMHB3ZU5abApmVUxqRmgwSVBob2k0TlB5L0pYYVVidWNycG45K2RWcDRHQUQKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNvakNDQWxTZ0F3SUJBZ0lVWmNiNE9WTmhhbXJOTU53dlRFT1J5RGFJMHhjd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE16TXpVeldoY05NelV3TnpFNU1UTXoKTkRJeldqQjdNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVNNQkFHQTFVRUN4TUpSR1ZqYTJodmRYTmxNU0l3CklBWURWUVFERXhsS1UwTWdSbXhoYm5RZ1NXNTBaWEp0WldScFlYUmxJRU5CTUNvd0JRWURLMlZ3QXlFQTYyeG4KcVJEUm5uSGZDRXJlNFliYW55aDNDT2p6Z3N1alFSenJGSlVtS2Ztamdld3dnZWt3RGdZRFZSMFBBUUgvQkFRRApBZ0VHTUE4R0ExVWRFd0VCL3dRRk1BTUJBZjh3SFFZRFZSME9CQllFRkJZaUI3bWp2cldPTUQxTVd5R0dWUjRiCkNjL3JNQjhHQTFVZEl3UVlNQmFBRkpCZ0s0NjYzVVdZMEhKbTVGb1plK1Rsem1mek1FY0dDQ3NHQVFVRkJ3RUIKQkRzd09UQTNCZ2dyQmdFRkJRY3dBb1lyYUhSMGNITTZMeTh4TWpjdU1DNHdMakU2T0RJd01DOTJNUzl3YTJrdApabXhoYm5RdGNtOXZkQzlqWVRBOUJnTlZIUjhFTmpBME1ES2dNS0F1aGl4b2RIUndjem92THpFeU55NHdMakF1Ck1UbzRNakF3TDNZeEwzQnJhUzFtYkdGdWRDMXliMjkwTDJOeWJEQUZCZ01yWlhBRFFRQnNOZWNUU0l6bmxvQWYKYkViU2JERlpGb20wMXc2WDN5WVpqVmFOQngrNEZmV3JNZ1ppU05rM3NkZ0dNRlpEVnVaZWl4OXlQUTk3bVIzWgpaTXU0Q3JRUAotLS0tLUVORCBDRVJUSUZJQ0FURS0tLS0t",
+			"LS0tLS1CRUdJTiBDRVJUSUZJQ0FURS0tLS0tCk1JSUNGVENDQWNlZ0F3SUJBZ0lVZElhblV0Nk0vNnFTL0RzMXJWSjZoc09JTmYwd0JRWURLMlZ3TUhneEN6QUoKQmdOVkJBWVRBbEpWTVE4d0RRWURWUVFJRXdaTmIzTmpiM2N4RHpBTkJnTlZCQWNUQmsxdmMyTnZkekVTTUJBRwpBMVVFQ2hNSlNsTkRJRVpzWVc1ME1SQXdEZ1lEVlFRTEV3ZEpibVp2YzJWak1TRXdId1lEVlFRREV4aEtVME1nClJteGhiblFnVW05dmRDQkRRU0JCTFRJd01qVXdIaGNOTWpVd056SXhNVE15TkRVMFdoY05ORFV3TnpFMk1UTXkKTlRJMFdqQjRNUXN3Q1FZRFZRUUdFd0pTVlRFUE1BMEdBMVVFQ0JNR1RXOXpZMjkzTVE4d0RRWURWUVFIRXdaTgpiM05qYjNjeEVqQVFCZ05WQkFvVENVcFRReUJHYkdGdWRERVFNQTRHQTFVRUN4TUhTVzVtYjNObFl6RWhNQjhHCkExVUVBeE1ZU2xORElFWnNZVzUwSUZKdmIzUWdRMEVnUVMweU1ESTFNQ293QlFZREsyVndBeUVBbGd1Q3hQRHUKV1VhajZjazZHdFIrZGdCNS9SRENSNmdhSWkzN3ZCQmx0ZE9qWXpCaE1BNEdBMVVkRHdFQi93UUVBd0lCQmpBUApCZ05WSFJNQkFmOEVCVEFEQVFIL01CMEdBMVVkRGdRV0JCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFmCkJnTlZIU01FR0RBV2dCU1FZQ3VPdXQxRm1OQnladVJhR1h2azVjNW44ekFGQmdNclpYQURRUUJxbTd5S21mVG0KV01PcFdHRldxSVJLcmJnRW44NkpyQmVWSXFXakVUMWs2NEVjR2pFN0JkdWxqdjdDVGhCQ2xwY1c3bzFuUGlVeApZeDFUNnMvSmU3b0kKLS0tLS1FTkQgQ0VSVElGSUNBVEUtLS0tLQo=",
+			time.Minute*11,
+			time.Minute*10,
+			time.Second*10,
+		),
+	)
+})
