+++ pkg/signver/hashivault/auth_jwt.go
@@ -1,27 +1,41 @@
 package hashivault
 
-import vault "github.com/hashicorp/vault/api"
+import (
+	"fmt"
+
+	vault "github.com/hashicorp/vault/api"
+)
 
 type jwtAuthenticator struct {
 	baseAuthenticator
-	jwtToken string
-	role     string
+	jwtProvider jwtTokenProvider
+	role        string
 }
 
-func newJWTAuthenticator(jwtToken, role string) *jwtAuthenticator {
+func newJWTAuthenticator(provider jwtTokenProvider, role string) *jwtAuthenticator {
 	return &jwtAuthenticator{
 		baseAuthenticator: baseAuthenticator{
 			authPath: "jwt",
 		},
-		jwtToken: jwtToken,
-		role:     role,
+		jwtProvider: provider,
+		role:        role,
 	}
 }
 
 func (j *jwtAuthenticator) Login(client *vault.Client) error {
+	if j.isTokenValid() {
+		client.SetToken(j.tokenID)
+		return nil
+	}
+
+	jwtToken, err := j.jwtProvider.GetToken()
+	if err != nil {
+		return fmt.Errorf("failed to get JWT token: %w", err)
+	}
+
 	loginData := map[string]interface{}{
 		"role": j.role,
-		"jwt":  j.jwtToken,
+		"jwt":  jwtToken,
 	}
 	return j.baseAuthenticator.login(client, loginData)
 }
